[MDC-199] twamp: bring down some datatype sizes
This will allow us to get rid of a lot of warnings.  But just in case,
add some static_assert() clauses to ensure int is large enough for our
needs.

static_assert() needs an include of assert.h, and it is C11.

+++ twamp-client-c/src/twamp.h
@@ -56,8 +56,8 @@ typedef struct twamp_parameters {
     unsigned int packets_count;
     unsigned int payload_size;
     unsigned int packets_max;
-    unsigned long int packets_interval_us;
-    unsigned long int packets_timeout_us;
+    unsigned int packets_interval_us;
+    unsigned int packets_timeout_us;
 } TWAMPParameters;
 
 /* Context */
